+++ src/components/forms/LoginForm.vue
@@ -11,9 +11,12 @@
 </template>
 
 <script lang="ts">
-import useValidate from '@vuelidate/core'
 import InputEmail from '@/components/form-elements/text-inputs/InputEmail.vue'
 import InputPassword from '@/components/form-elements/text-inputs/InputPassword.vue'
+import userLogin from '@/services/apiMethods/userLogin'
+import checkUserExist from '@/services/apiMethods/checkUserExist'
+import { TIMEOUT_REDIRECT } from '@/constants/projectConfigs'
+import { useUserStore } from '@/stores/User'
 
 export default {
   name: 'LoginForm',
@@ -21,12 +24,6 @@ export default {
   components: {
     InputEmail,
     InputPassword
-  },
-
-  setup() {
-    return {
-      v$: useValidate()
-    }
   },
 
   data() {
@@ -39,21 +36,46 @@ export default {
   },
 
   methods: {
-    submitLoginForm() {
-      this.$emit('loginAction', {
-        email: this.loginForm.email,
-        password: this.loginForm.password
-      })
+    async submitLoginForm() {
+      try {
+        const doesUserExist = await checkUserExist(this.loginForm.email)
+        if (doesUserExist) {
+          const loginResult = await userLogin(this.loginForm.email, this.loginForm.password)
+          if (loginResult.statusCode === 200) {
+            this.$emit(
+              'successMessage',
+              'Вы успешно вошли в учетную запись.'
+            )
+            const user = useUserStore()
+            user.login()
+            user.setUserMail(this.loginForm.email)
+            setTimeout(() => {
+              this.$router.push('/')
+            }, TIMEOUT_REDIRECT)
+          } else if (loginResult.statusCode === 400) {
+            this.$emit(
+              'errorMessage',
+              'Неверный пароль. Пожалуйста, исправьте введенный пароль и попробуйте еще раз.'
+            )
+          } else {
+            this.$emit(
+              'errorMessage',
+              'Что-то пошло не так. Повторите попытку позже.'
+            )
+          }
+        }
+      } catch (error: unknown) {
+        this.$emit(
+          'errorMessage',
+          'Что-то пошло не так. Повторите попытку позже.'
+        )
+      }
     }
   }
 }
 </script>
 
 <style lang="scss" scoped>
-@import '@/assets/styles/variables';
-@import '@/assets/styles/mixins';
-@import '@/assets/styles/extends';
-
 .login-form {
   margin-top: 50px;
 }
